Implement bondholder income calculator (#17)
* Implement bondholder income calculator

* add tests for bondholder_income_calculator service, fix func doc string,
replace Bond properties types to Decimal. Fix all tests (make them async
if needed only, not by default)

* Some linter adjustments\fixes (some kind of QoL)

* fix pytest warnings, adjust tests

* remove unused package from req

* income calculator implementation.

* Move use case deps to separate directory, add month income calculations router, deps and use case

* add month income calculation into the BondModal

* add tests for calculate_income use case

+++ src/application/dto/bond.py
@@ -1,33 +1,34 @@
 from dataclasses import dataclass
 from uuid import UUID
+from decimal import Decimal
 
 
 @dataclass
 class BondDTO:
     id: UUID
-    nominal_value: float
+    nominal_value: Decimal
     series: str
     maturity_period: int
-    initial_interest_rate: float
+    initial_interest_rate: Decimal
     first_interest_period: int
-    reference_rate_margin: float
+    reference_rate_margin: Decimal
 
 
 @dataclass
 class BondCreateDTO:
     series: str
-    nominal_value: float
+    nominal_value: Decimal
     maturity_period: int
-    initial_interest_rate: float
+    initial_interest_rate: Decimal
     first_interest_period: int
-    reference_rate_margin: float
+    reference_rate_margin: Decimal
 
 
 @dataclass
 class BondUpdateDTO:
-    nominal_value: float | None = None
+    nominal_value: Decimal | None = None
     series: str | None = None
     maturity_period: int | None = None
-    initial_interest_rate: float | None = None
+    initial_interest_rate: Decimal | None = None
     first_interest_period: int | None = None
-    reference_rate_margin: float | None = None
+    reference_rate_margin: Decimal | None = None
