+++ src/application/dto/bondholder.py
@@ -1,5 +1,6 @@
 from dataclasses import dataclass
 from datetime import datetime, date
+from decimal import Decimal
 from uuid import UUID
 
 
@@ -11,11 +12,11 @@ class BondHolderDTO:
     bond_id: UUID
 
     series: str
-    nominal_value: float
+    nominal_value: Decimal
     maturity_period: int
-    initial_interest_rate: float
+    initial_interest_rate: Decimal
     first_interest_period: int
-    reference_rate_margin: float
+    reference_rate_margin: Decimal
 
     id: UUID | None = None
     last_update: datetime | None = None
